requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
cryptography
argon2-cffi
python-dotenv
pyinstaller
